backend/requirements.txt
mtcnn==0.1.1
tensorflow==2.4.1
Flask
opencv-python==4.5.5.62
tensorflow_addons
Pillow
numpy
pillow
matplotlib